Updated Html Content
I updated the html this weekend. I removed the <br> from the markup for the forms as this can be done with CSS.  I swapped the position of the form markup and the buttons Cid created from <li> elements for layout purposes.  I added 2 wrapper classes again for layout functionality.