Updated styles.css
I created a 2 column layout that needs a little bit more formatting.  I increased the size of the text page wide.  Added focus effects for the form as well as formatting the form layout.  Implemented "display:block;" for the label selector removing the need for html <br> element in the markup. Increased the size of the border and formatted the text/placeholder within the input fields.  Added "title" attributes to the buttons Cid created providing more information to the user on mouse over.