Fix the tab alignment.
The tabs weren't aligning correctly because of using `display: inline` rather than `display: inline-block`.

When an element is `display: inline`, the values for padding, margin, and borders do not contribute to the element's height. It will always only treated as being as tall as its content when the browser is laying out the elements.

Changing to `display: inline-block` permits the padding on the `li` element to be included in the element's height and allows it to lay out correctly.